+++ contrast-finder-webapp/src/main/webapp/Js/36-sample.color.js
@@ -1,59 +1,46 @@
 $(document).ready(function() {
-    changeBackgroundSample();
-    changeForegroundSample();
+    // changeColorSample("foreground");
+    // changeColorSample("background");
     document.getElementById("foreground-input").onchange = function() {
-        changeForegroundSample();
-        console.log("change foreground Appel");
+        changeColorSample("foreground");
     };
     document.getElementById("background-input").onchange = function() {
-        changeBackgroundSample();
+        changeColorSample("background");
     };
 });
 
-function changeForegroundSample() {
-    var input = document.getElementById("foreground-input");
-    var color = input.value;
-    var sample = document.getElementById("foreground-sample");
-    color = isValidateColor(color.toString());
-    if (color !== "false") {
-        sample.style.backgroundColor = color;
-        sample.classList.add('color-sample');
-        sample.classList.add('sample-bordered');
-        document.getElementById("foreground-sample-invalid").style.display = "none";
-        input.classList.remove('error');
-    } else {
-        sample.style.backgroundColor = "rgba(0,0,0,0)";
-        sample.classList.remove('color-sample');
-        sample.classList.remove('sample-bordered');
-        document.getElementById("foreground-sample-invalid").style.display = "inherit";
-        input.classList.add('error');
-    }
-}
 
-function changeBackgroundSample() {
-    var input = document.getElementById("background-input");
-    var color = input.value;
-    var sample = document.getElementById("background-sample");
+function changeColorSample(colorPrefix) {
+    var input = document.getElementById(colorPrefix + "-input");
+    var color = input.value.toLowerCase();
+    var sample = document.getElementById(colorPrefix + "-sample");
+    color = color.toString().replace(/\s/g,""); // replace ' ', \t, \n, ...
     color = isValidateColor(color.toString());
     if (color !== "false") {
         sample.style.backgroundColor = color;
         sample.classList.add('color-sample');
         sample.classList.add('sample-bordered');
-        document.getElementById("background-sample-invalid").style.display = "none";
+        document.getElementById(colorPrefix + "-sample-invalid").style.display = "none";
         input.classList.remove('error');
     } else {
         sample.style.backgroundColor = "rgba(0,0,0,0)";
         sample.classList.remove('color-sample');
         sample.classList.remove('sample-bordered');
-        document.getElementById("background-sample-invalid").style.display = "inherit";
+        document.getElementById(colorPrefix + "-sample-invalid").style.display = "inherit";
         input.classList.add('error');
     }
 }
 
 function isValidateColor(str) {
-    if (str.match(/^#[a-fA-F0-9]{6}$/i) !== null
-            || str.match(/^#[a-fA-F0-9]{3}$/i) !== null)
-    {
+    str = str.trim();
+    if(str.match(/^rgb\((\d{1,3}),(\d{1,3}),(\d{1,3})\)$/) !== null){
+        return str;
+    }
+    else if (str.match(/^(\d{1,3}),(\d{1,3}),(\d{1,3})$/) !== null){
+        return "rgb("+ str + ")";
+    }
+    else if (str.match(/^#[a-fA-F0-9]{6}$/i) !== null
+            || str.match(/^#[a-fA-F0-9]{3}$/i) !== null){
         return str;
     }
     else if (str.match(/^#?([a-fA-F0-9]{6})$/) !== null
@@ -62,6 +49,20 @@ function isValidateColor(str) {
         return str;
     }
     else {
+        // source: https://gist.github.com/bobspace/2712980
+        var CssColorNames = ["AliceBlue","AntiqueWhite","Aqua","Aquamarine","Azure","Beige","Bisque","Black","BlanchedAlmond","Blue","BlueViolet","Brown","BurlyWood","CadetBlue","Chartreuse","Chocolate","Coral","CornflowerBlue","Cornsilk","Crimson","Cyan","DarkBlue","DarkCyan","DarkGoldenRod","DarkGray","DarkGrey","DarkGreen","DarkKhaki","DarkMagenta","DarkOliveGreen","Darkorange","DarkOrchid","DarkRed","DarkSalmon","DarkSeaGreen","DarkSlateBlue","DarkSlateGray","DarkSlateGrey","DarkTurquoise","DarkViolet","DeepPink","DeepSkyBlue","DimGray","DimGrey","DodgerBlue","FireBrick","FloralWhite","ForestGreen","Fuchsia","Gainsboro","GhostWhite","Gold","GoldenRod","Gray","Grey","Green","GreenYellow","HoneyDew","HotPink","IndianRed","Indigo","Ivory","Khaki","Lavender","LavenderBlush","LawnGreen","LemonChiffon","LightBlue","LightCoral","LightCyan","LightGoldenRodYellow","LightGray","LightGrey","LightGreen","LightPink","LightSalmon","LightSeaGreen","LightSkyBlue","LightSlateGray","LightSlateGrey","LightSteelBlue","LightYellow","Lime","LimeGreen","Linen","Magenta","Maroon","MediumAquaMarine","MediumBlue","MediumOrchid","MediumPurple","MediumSeaGreen","MediumSlateBlue","MediumSpringGreen","MediumTurquoise","MediumVioletRed","MidnightBlue","MintCream","MistyRose","Moccasin","NavajoWhite","Navy","OldLace","Olive","OliveDrab","Orange","OrangeRed","Orchid","PaleGoldenRod","PaleGreen","PaleTurquoise","PaleVioletRed","PapayaWhip","PeachPuff","Peru","Pink","Plum","PowderBlue","Purple","Red","RosyBrown","RoyalBlue","SaddleBrown","Salmon","SandyBrown","SeaGreen","SeaShell","Sienna","Silver","SkyBlue","SlateBlue","SlateGray","SlateGrey","Snow","SpringGreen","SteelBlue","Tan","Teal","Thistle","Tomato","Turquoise","Violet","Wheat","White",
+                             "WhiteSmoke","Yellow","YellowGreen"];
+        var query     = str.toLowerCase();
+        var colorName = "";
+        CssColorNames.some(function(element, i) {
+            if (query === element.toLowerCase()) {
+                colorName = element;
+                return true;
+            }
+        });
+        if(colorName !== ""){
+            return colorName;
+        }
         return "false"
     }
 }